refactor: drop redundant fdv2_ prefix from filenames
Files inside the fdv2/ directory don't need the fdv2_ prefix.
Renames: fdv2_payload.dart → payload.dart, fdv2_protocol_handler.dart
→ protocol_handler.dart, fdv2_protocol_types.dart → protocol_types.dart,
fdv2_source_result.dart → source_result.dart,
fdv2_flag_eval_mapper.dart → flag_eval_mapper.dart. Test files follow
the same pattern.
