+++ packages/common_client/lib/src/data_sources/fdv2/fdv2.dart
@@ -1,9 +1,9 @@
 /// FDv2 protocol and data model types.
 library;
 
-export 'fdv2_flag_eval_mapper.dart';
-export 'fdv2_payload.dart';
-export 'fdv2_protocol_handler.dart';
-export 'fdv2_protocol_types.dart';
-export 'fdv2_source_result.dart';
+export 'flag_eval_mapper.dart';
+export 'payload.dart';
+export 'protocol_handler.dart';
+export 'protocol_types.dart';
+export 'source_result.dart';
 export 'selector.dart';
